requirements.txt
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==3.0.3
django-activity-stream==0.9.0
django-appconf==1.0.4
django-cors-headers==3.2.1
django-imagekit==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
httpie==2.1.0
idna==2.9
inflection==0.4.0
install==1.3.3
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.3.14
openapi-codec==1.3.2
packaging==20.3
pilkit==2.0
Pillow==7.2.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
PyMySQL==0.9.3
pyparsing==2.4.7
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wincertstore==0.2
wrapt==1.11.2
